Move some things arround some more and update rest of the project with new structure
