+++ msftbandlib-tests/MSFTBandLib.test.cs
@@ -1,6 +1,5 @@
 using MSFTBandLib;
 using MSFTBandLib.Command;
-using MSFTBandLib.Facility;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using MSFTBandLib.Contracts.Types;
