+++ msftbandlib.Contracts/msftbandlib.Contracts.csproj
@@ -6,4 +6,12 @@
     <RootNamespace>MSFTBandLib.Contracts</RootNamespace>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Atr\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\msftbandlib.Shared\msftbandlib.Shared.csproj" />
+  </ItemGroup>
+
 </Project>
