+++ msftbandlib.Shared/msftbandlib.Shared.csproj
@@ -1,0 +1,9 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard1.4</TargetFramework>
+    <AssemblyName>MSFTBandLib.Shared</AssemblyName>
+    <RootNamespace>MSFTBandLib.Shared</RootNamespace>
+  </PropertyGroup>
+
+</Project>
