+++ msftbandlib.sln
@@ -16,6 +16,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "msftbandlib-dotnetcore", "m
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "msftbandlib.Contracts", "msftbandlib.Contracts\msftbandlib.Contracts.csproj", "{E488C536-4F0F-4477-83A2-B511A655330F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "msftbandlib.Shared", "msftbandlib.Shared\msftbandlib.Shared.csproj", "{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -130,6 +132,26 @@ Global
 		{E488C536-4F0F-4477-83A2-B511A655330F}.Release|x64.Build.0 = Release|Any CPU
 		{E488C536-4F0F-4477-83A2-B511A655330F}.Release|x86.ActiveCfg = Release|Any CPU
 		{E488C536-4F0F-4477-83A2-B511A655330F}.Release|x86.Build.0 = Release|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Debug|ARM.Build.0 = Debug|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Debug|x64.Build.0 = Debug|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Debug|x86.Build.0 = Debug|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Release|ARM.ActiveCfg = Release|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Release|ARM.Build.0 = Release|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Release|ARM64.Build.0 = Release|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Release|x64.ActiveCfg = Release|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Release|x64.Build.0 = Release|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Release|x86.ActiveCfg = Release|Any CPU
+		{C1F9A62A-E2A7-47FF-B536-DEAA7B8EBE2E}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
